added consequent_differs and a test for it
